+++ src/base/type/physicalvalue.h
@@ -1,27 +1,33 @@
 #ifndef TYPE_PHYSICALVALUE
 #define TYPE_PHYSICALVALUE
 
-#include <cstdint>
+#include "base/conv/parse.h"
+#include "base/conv/tostring.h"
+#include "base/refl/auto_enum.h"
+#include "base/text/valueunit.h"
 
 namespace Type
 {
 
-enum class SimpleUnit : std::uint8_t { none, relative, absolute };
-
-template <typename Value, typename Unit = SimpleUnit>
-class PhysicalValue
+template <typename Value, Refl::is_enum Unit> struct PhysicalValue
 {
-public:
-	Value value;
-	Unit unit;
-
-	constexpr PhysicalValue() : value{}, unit{} {}
-	constexpr PhysicalValue(Value value, Unit unit) :
-	    value(value),
-	    unit(unit)
-	{}
+	Value value{};
+	Unit unit{};
 
 	constexpr bool operator==(const PhysicalValue &) const = default;
+
+	template <std::same_as<double> = Value>
+	static PhysicalValue fromString(const std::string &str)
+	{
+		Text::ValueUnit vu{str};
+		return {vu.getValue(), Refl::get_enum<Unit>(vu.getUnit())};
+	}
+
+	std::string toString() const
+	{
+		return Conv::toString(value)
+		     + std::string{Conv::toString(unit)};
+	}
 };
 
 }